+++ packages/vendure-plugin-store-credit/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Accept amount as input to the store credit payment handler
 - Export helper functions to populate wallets for customers
+- BREAKING: DB migration required: Customers can not have multiple wallets with the same name
 
 # 1.0.0 (2026-02-03)